requirements-minimal.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# PDF processing
PyMuPDF==1.23.8
PyPDF2==3.0.1
reportlab==4.0.7

# OpenAI integration (primary processing method - 92% accuracy)
openai>=1.0.0
pdf2image>=1.16.0
Pillow>=9.0.0

# Essential dependencies only
pytesseract==0.3.10
numpy>=1.21.0
requests>=2.28.0

# Production server
gunicorn==21.2.0

# Authentication and security
sqlalchemy==2.0.23
bcrypt==4.1.2
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
pyotp==2.9.0
qrcode[pil]==7.4.2
email-validator==2.1.0